# server/handlers/video.go
package handlers

import (
	"os"
	"path/filepath"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/nil0j/jirafeitor/config"
	"github.com/nil0j/jirafeitor/models/responses"
	"github.com/nil0j/jirafeitor/repository"
)

// @Tags Videos
// @Param id path int true "Video ID"
// @Success 200 {file} file "Video file"
// @Success 400 {object} responses.jsonError
// @Router /video/source/{id} [get]
func GetVideo(c *gin.Context) {
	folderPath := config.Data.Filesystem + c.Param("id")
	files, err := os.ReadDir(folderPath)
	if err != nil {
		responses.HandleError(c, err)
		return
	}

	if len(files) == 0 {
		responses.HandleError(c, err)
		return
	}

	var targetFileName string
	if filepath.Ext(files[0].Name()) == ".png" {
		targetFileName = files[1].Name()
	} else {
		targetFileName = files[0].Name()
	}

	filePath := filepath.Join(folderPath, targetFileName)
	c.File(filePath)
}

// @Tags Videos
// @Param id path int true "Video ID"
// @Success 200 {file} file "Image file"
// @Success 400 {object} responses.jsonError
// @Router /video/thumbnail/{id} [get]
func GetVideoThumbnail(c *gin.Context) {
	folderPath := config.Data.Filesystem + c.Param("id")
	files, err := os.ReadDir(folderPath)
	if err != nil {
		responses.HandleError(c, err)
		return
	}

	if len(files) == 0 {
		responses.HandleError(c, err)
		return
	}

	filePath := filepath.Join(folderPath, "thumbnail.png")
	c.File(filePath)
}

// @Tags Videos
// @Param id path int true "Video ID"
// @Success 200 {object} postgres.Video
// @Success 400 {object} responses.jsonError
// @Router /video/info/{id} [get]
func GetVideoInfo(c *gin.Context) {
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		responses.HandleError(c, err)
		return
	}

	videoInfo, err := repository.GetVideoInfo(id)
	if err != nil {
		responses.HandleError(c, err)
		return
	}

	responses.HandleVideoInfo(c, videoInfo)
}

// @Tags Videos
// @Param limit query int false "Limit amount of videos"
// @Success 200 {object} []postgres.Video
// @Success 400 {object} responses.jsonError
// @Router /videos [get]
func GetRecentVideos(c *gin.Context) {
	var limit int
	{
		limitInput := c.Query("limit")

		var err error
		limit, err = strconv.Atoi(limitInput)
		if err != nil {
			limit = 10 // default limit
		}
	}

	videos, err := repository.GetRecentVideos(limit)
	if err != nil {
		responses.HandleError(c, err)
		return
	}

	c.JSON(200, videos)
}